// frontend/src/api/esquema.d.ts
export interface components {
    schemas: {
// ...
        LicitacaoDetalhe: {
            /** Id */
            id: number;
            /** Fonte */
            fonte: string;
            /** Numero */
            numero: string | null;
            /** Objeto */
            objeto: string | null;
            /** Modalidade */
            modalidade: string | null;
            /** Orgao */
            orgao: string | null;
            /** Abertura */
            abertura: string | null;
            /**
             * Situacao
             * @enum {string}
             */
            situacao: "aberta" | "andamento" | "suspensa" | "encerrada" | "cancelada" | "desconhecida";
            /** Rotulo */
            rotulo: string | null;
            /** Dias Para Abertura */
            dias_para_abertura: number | null;
            /** Publicacao */
            publicacao: string | null;
            /** Criterio */
            criterio: string | null;
            /** Homologacao */
            homologacao: string | null;
            link_origem?: components["schemas"]["LinkOrigem"] | null;
            /** Fases */
            fases: components["schemas"]["Fase"][];
            /** Contratos */
            contratos: components["schemas"]["Contrato"][];
            /** Vencedores */
            vencedores: components["schemas"]["Vencedor"][];
            /** Itens */
            itens: components["schemas"]["Item"][];
            /** Arquivos */
            arquivos: components["schemas"]["Arquivo"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}